+++ lib/material_drum_picker.dart
@@ -16,6 +16,8 @@ export 'src/theme/drum_picker_theme.dart';
 export 'src/utils/drum_date_utils.dart' show DrumDateUtils;
 
 // Widgets
+export 'src/widgets/drum_date_time_picker_dialog.dart'
+    show showDrumDateTimePicker;
 export 'src/widgets/drum_picker.dart'
     show DrumPicker, DrumSelectableDayPredicate;
 export 'src/widgets/drum_picker_dialog.dart' show showDrumDatePicker;